+++ packages/plugin-packages/mashroom-cdn/type-definitions/index.d.ts
@@ -1,0 +1,7 @@
+
+export interface MashroomCDNService {
+    /**
+     * Return a CDN host or undefined if there is none configured.
+     */
+    getCDNHost(): string | undefined;
+}