+++ packages/plugin-packages/mashroom-cdn/type-definitions/index.js
@@ -1,0 +1,8 @@
+// @flow
+
+export interface MashroomCDNService {
+    /**
+     * Return a CDN host or undefined if there is none configured.
+     */
+    getCDNHost(): ?string;
+}